requirements.txt
flask==3.0.0
flask-cors==4.0.0
flask-jwt-extended==4.5.3
werkzeug==3.0.1
pypdf2==3.0.1
requests==2.31.0
feedparser==6.0.10
flask-sqlalchemy==3.1.1
python-docx==1.1.0
spacy==3.7.2
google-genai==0.3.0
